requirements/local.txt
-r ./common.txt

django-debug-toolbar==3.2.2
freezegun==1.1.0
black==21.6b0
factory_boy==3.2.0
psycopg2-binary==2.9.1
pytest==6.2.4
pytest-django==4.4.0